Make splice/filter operation clearer
The array has to be spread before filtering.
This is for the case where only a later image has a defined width/height
In those cases the width/height array would be defined as [4: 350]
for example.
Filtering on that array would give unexpected result.
The spreading creates the full array and then filters correctly.